+++ libnodegl/node_block.c
@@ -89,6 +89,7 @@ static const struct param_choices layout_choices = {
                                        NGL_NODE_ANIMATEDVEC3,           \
                                        NGL_NODE_ANIMATEDVEC4,           \
                                        NGL_NODE_ANIMATEDQUAT,           \
+                                       NGL_NODE_ANIMATEDCOLOR,          \
                                        NGL_NODE_STREAMEDINT,            \
                                        NGL_NODE_STREAMEDIVEC2,          \
                                        NGL_NODE_STREAMEDIVEC3,          \